Store compression and compression level in header class

+++ Autoupdate/SUBinaryDeltaCreate.h
@@ -13,6 +13,6 @@
 #import "SPUDeltaArchiveProtocol.h"
 
 @class NSString;
-BOOL createBinaryDelta(NSString *source, NSString *destination, NSString *patchFile, SUBinaryDeltaMajorVersion majorVersion, SPUDeltaCompressionMode compression, int32_t compressionLevel, BOOL verbose, NSError * __autoreleasing *error);
+BOOL createBinaryDelta(NSString *source, NSString *destination, NSString *patchFile, SUBinaryDeltaMajorVersion majorVersion, SPUDeltaCompressionMode compression, uint8_t compressionLevel, BOOL verbose, NSError * __autoreleasing *error);
 
 #endif
